fix: resolve PWA context lint warning and enforce zero warnings
- Extract PWAContext to separate file (pwa-context-value.ts) to fix
  react-refresh/only-export-components warning
- Add --max-warnings 0 to lint script to prevent future regressions
